Core: Move variable setup and derive to BinaryBH example
The variableSetupUp() should be defined in the specific example level
class as it needs to be static and it sets up the derived variables.
