+++ Examples/BinaryBH/BinaryBHLevel.hpp
@@ -14,6 +14,7 @@
 class BinaryBHLevel : public GRAMRLevel
 {
   public:
+    static void variableSetUp();
 
     // Inherit the contructors from GRAMRLevel
     using GRAMRLevel::GRAMRLevel;
@@ -37,6 +38,14 @@ class BinaryBHLevel : public GRAMRLevel
 
     void errorEst(amrex::TagBoxArray &tag_box_array, int clearval, int tagval,
                   amrex::Real time, int n_error_buf = 0, int ngrow = 0) final;
+
+    //! Calculate derived quantity "name" and store in multifab starting at
+    //! dcomp
+    void derive(const std::string &name, amrex::Real time,
+                amrex::MultiFab &multifab, int dcomp) final;
+
+    // NOLINTNEXTLINE(cppcoreguidelines-avoid-non-const-global-variables)
+    static amrex::Vector<std::string> plot_constraints;
 
 #ifdef AMREX_USE_HDF5
     /// Any actions that should happen just before plot files output
