+++ Source/GRTeclynCore/GRAMRLevel.cpp
@@ -5,149 +5,6 @@
 
 #include "GRAMRLevel.hpp"
 
-// NOLINTNEXTLINE(cppcoreguidelines-avoid-non-const-global-variables)
-amrex::Vector<std::string> GRAMRLevel::plot_constraints;
-
-struct GRAMRBCFill
-{
-    AMREX_GPU_DEVICE
-    void operator()(const amrex::IntVect & /*iv*/,
-                    const amrex::Array4<amrex::Real> & /*dest*/,
-                    const int /*dcomp*/, const int /*numcomp*/,
-                    const amrex::GeometryData & /*geom*/,
-                    const amrex::Real /*time*/, const amrex::BCRec * /*bcr*/,
-                    const int /*bcomp*/, const int /*orig_comp*/) const
-    {
-        // We don't have any external Dirichlet BC
-    }
-};
-
-void gramr_bc_fill(const amrex::Box &box, amrex::FArrayBox &data,
-                   const int dcomp, const int numcomp,
-                   const amrex::Geometry &geom, const amrex::Real time,
-                   const amrex::Vector<amrex::BCRec> &bcr, const int bcomp,
-                   const int scomp)
-{
-    amrex::GpuBndryFuncFab<GRAMRBCFill> bndry_func(GRAMRBCFill{});
-    bndry_func(box, data, dcomp, numcomp, geom, time, bcr, bcomp, scomp);
-}
-
-void GRAMRLevel::variableSetUp()
-{
-    BL_PROFILE("GRAMRLevel::variableSetUp()");
-    const int nghost = simParams().num_ghosts;
-    desc_lst.addDescriptor(State_Type, amrex::IndexType::TheCellType(),
-                           amrex::StateDescriptor::Point, nghost, NUM_VARS,
-                           &amrex::cell_quartic_interp);
-
-    BoundaryConditions::params_t bparms = simParams().boundary_params;
-    BoundaryConditions boundary_conditions;
-    boundary_conditions.define(simParams().center, bparms,
-                               amrex::DefaultGeometry(), nghost);
-
-    amrex::Vector<amrex::BCRec> bcs(NUM_VARS);
-    for (int icomp = 0; icomp < NUM_VARS; ++icomp)
-    {
-        auto &bc = bcs[icomp];
-        for (amrex::OrientationIter oit; oit.isValid(); ++oit)
-        {
-            amrex::Orientation face = oit();
-            const int idim          = face.coordDir();
-            const int bctype = boundary_conditions.get_boundary_condition(face);
-            if (amrex::DefaultGeometry().isPeriodic(idim))
-            {
-                bc.set(face, amrex::BCType::int_dir);
-            }
-            else if (bctype == BoundaryConditions::STATIC_BC ||
-                     bctype == BoundaryConditions::SOMMERFELD_BC ||
-                     bctype == BoundaryConditions::MIXED_BC)
-            {
-                bc.set(face, amrex::BCType::foextrap);
-            }
-            else if (bctype == BoundaryConditions::REFLECTIVE_BC)
-            {
-                int parity = boundary_conditions.get_var_parity(
-                    icomp, idim, VariableType::evolution);
-                if (parity == 1)
-                {
-                    bc.set(face, amrex::BCType::reflect_even);
-                }
-                else
-                {
-                    bc.set(face, amrex::BCType::reflect_odd);
-                }
-            }
-            else if (bctype == BoundaryConditions::EXTRAPOLATING_BC)
-            {
-                amrex::Abort("xxxxx EXTRAPOLATING_BC todo");
-            }
-            else
-            {
-                amrex::Abort("Unknow BC type " + std::to_string(bctype));
-            }
-        }
-    }
-
-    amrex::Vector<std::string> name(NUM_VARS);
-    for (int i = 0; i < NUM_VARS; ++i)
-    {
-        name[i] = UserVariables::variable_names[i];
-    }
-
-    amrex::StateDescriptor::BndryFunc bndryfunc(gramr_bc_fill);
-    bndryfunc.setRunOnGPU(true); // Run the bc function on gpu.
-
-    desc_lst.setComponent(State_Type, 0, name, bcs, bndryfunc);
-
-    amrex::ParmParse pp("amr");
-    if (pp.contains("derive_plot_vars"))
-    {
-        std::vector<std::string> names;
-        pp.getarr("derive_plot_vars", names);
-
-        // Constraints
-        auto names_it =
-            std::find_if(names.begin(), names.end(),
-                         [](const std::string &name) {
-                             return std::string("ham") == amrex::toLower(name);
-                         });
-        if (names_it != names.end())
-        {
-            names.erase(names_it);
-            plot_constraints.push_back("Ham");
-        }
-        //
-        names_it =
-            std::find_if(names.begin(), names.end(),
-                         [](const std::string &name) {
-                             return std::string("mom") == amrex::toLower(name);
-                         });
-        if (names_it != names.end())
-        {
-            names.erase(names_it);
-            plot_constraints.push_back("Mom1");
-            plot_constraints.push_back("Mom2");
-            plot_constraints.push_back("Mom3");
-        }
-        //
-        if (!plot_constraints.empty())
-        {
-            names.emplace_back("constraints");
-            pp.addarr("derive_plot_vars", names);
-
-            derive_lst.add(
-                "constraints", amrex::IndexType::TheCellType(),
-                static_cast<int>(plot_constraints.size()), plot_constraints,
-                amrex::DeriveFuncFab(), // null function because we won't use
-                                        // it.
-                [=](const amrex::Box &box) { return amrex::grow(box, nghost); },
-                &amrex::cell_quartic_interp);
-            derive_lst.addComponent("constraints", desc_lst, State_Type, 0,
-                                    NUM_VARS);
-        }
-    }
-}
-
 void GRAMRLevel::variableCleanUp()
 {
     desc_lst.clear();
@@ -339,17 +196,37 @@ void GRAMRLevel::writePlotFilePre(const std::string & /*dir*/,
                                   std::ostream & /*os*/)
 {
     m_is_writing_plotfile = true;
-    if (!plot_constraints.empty())
-    {
-        auto &state_new = get_new_data(State_Type);
-        FillPatch(*this, state_new, state_new.nGrow(),
-                  get_state_data(State_Type).curTime(), State_Type, 0,
-                  state_new.nComp()); // xxxxx Do we need all components?
-    }
 }
 
 void GRAMRLevel::writePlotFilePost(const std::string & /*dir*/,
                                    std::ostream & /*os*/)
 {
     m_is_writing_plotfile = false;
+}
+
+// NOLINTBEGIN(bugprone-easily-swappable-parameters)
+std::unique_ptr<amrex::MultiFab> GRAMRLevel::derive(const std::string &name,
+                                                    amrex::Real time, int ngrow)
+// NOLINTEND(bugprone-easily-swappable-parameters)
+{
+    std::unique_ptr<amrex::MultiFab> multifab;
+    const amrex::DeriveRec *rec = derive_lst.get(name);
+    if (rec != nullptr)
+    {
+        multifab = std::make_unique<amrex::MultiFab>(
+            this->boxArray(), this->DistributionMap(), rec->numState(), ngrow);
+        derive(name, time, *multifab, 0);
+    }
+    else
+    {
+        amrex::Abort("Unknown derived variable");
+    }
+    return multifab;
+}
+
+void GRAMRLevel::derive(const std::string &name, amrex::Real time,
+                        amrex::MultiFab &multifab, int dcomp)
+{
+    amrex::Abort("GRAMRLevel::derive(): Implement this function in the child "
+                 "level class if derived variables are required.");
 }
